changed type

+++ cpp04/ex01/main.cpp
@@ -17,19 +17,19 @@ int main(){
 
 	//Own tests
 	std::cout << "\n\033[33m---------Extra Tests---------\033[0m\n";
-	int amount = 10;
+	size_t amount = 10;
 	Animal* zoo[amount];
-	for (int i = 0; i < amount; i++){	
+	for (size_t i = 0; i < amount; i++){	
 		if (i < (amount / 2))
 			zoo[i] = new Dog();
 		else
 			zoo[i] = new Cat();
 	}
-	for (int i = 0; i < amount; i++){
+	for (size_t i = 0; i < amount; i++){
 		zoo[i]->makeSound();
 	}
 	std::cout << "\033[31m-----Calling destructors-----\033[0m\n";
-	for (int i = 0; i < amount; i ++)
+	for (size_t i = 0; i < amount; i ++)
 		delete zoo[i] ;
 	return 0;
 }
